Refactor CommandParsing
- Move CommandSet annotation to CommandDSL
- Move produceContainer to CommandDSL
- Rename getCommandStruct to cleanCommandMessage
- Remove unnecessary parentheses
