+++ src/main/kotlin/me/aberrantfox/hotbot/commandframework/commands/DataCommands.kt
@@ -1,7 +1,7 @@
 package me.aberrantfox.hotbot.commandframework.commands
 
 import me.aberrantfox.hotbot.commandframework.ArgumentType
-import me.aberrantfox.hotbot.commandframework.CommandSet
+import me.aberrantfox.hotbot.dsls.command.CommandSet
 import me.aberrantfox.hotbot.dsls.command.commands
 import me.aberrantfox.hotbot.dsls.embed.embed
 import me.aberrantfox.hotbot.extensions.jda.fullName
