+++ src/main/kotlin/me/aberrantfox/hotbot/commandframework/commands/UtilityCommands.kt
@@ -2,7 +2,7 @@ package me.aberrantfox.hotbot.commandframework.commands
 
 import com.google.gson.Gson
 import me.aberrantfox.hotbot.commandframework.ArgumentType
-import me.aberrantfox.hotbot.commandframework.CommandSet
+import me.aberrantfox.hotbot.dsls.command.CommandSet
 import me.aberrantfox.hotbot.database.savePermissions
 import me.aberrantfox.hotbot.dsls.command.CommandEvent
 import me.aberrantfox.hotbot.services.saveConfig
